requirements.txt
vk_api
pandas
matplotlib
weasyprint
python-dotenv
numpy